ppp: Add ppp
Add a subdirectory with the files required to build the ppp base image.
Create also a basic README in the root directory that introduces the
available images (only ppp so far). Add a recursive Makefile structure
which enables a build process for several images in the
ppp-pppoe-toolkit.

The ppp image is based on alpine and installs the ppp package. The bash
package is installed for convenience. The Dockerfile of ppp also creates
a symlink from `/var/log/ppp/pppd.stdout.log` to `/proc/1/fd/1` which
can be used to log the pppd output to stdout of the process with PID 1.
Therefore, the output will appear in docker logs. The files
`ppp/etc/ppp/options` and `ppp/etc/ppp/default` provide some reasonable
default args for the usage of the image.