rp-pppoe: Add rp-pppoe
Add a new subproject with the required files for the rp-pppoe image
build. Create a basic README with usage examples, that demonstrate
session establishment between pppd and pppoe-server. Add a Makefile for
the recursive image build process.

The rp-pppoe image adds the `rp-pppoe` package on top of the ppp image.
It also provides a `pppoe-server-options` configuration file that works
with the pppd client configuration in the ppp image out-of-the-box.  The
`/var/log/ppp/rp-pppoe-pppd.stdout.log` logfile is symlinked to
`/proc/1/fd/1` analog to the pppd log in ppp, to redirect the pppd
output from pppoe-server (the server uses pppd under the hood) to the
stdout of the process with PID 1. Therefore, logs will be available
through the docker logs command.