ppp: Install ifupdown-ng
Install the ifupdown-ng package in the ppp image. Add a docker
entrypoint script that allows the execution of interface configurations
on container startup. Add an example to the README that demonstrates how
an example VLAN configuration can be applied with ifupdown-ng on
startup.

ifupdown-ng is a network device manager that allows you to apply network
device configurations through `interfaces` files.